[feat] 불필요한 interface 삭제
